Split ChatBar.tsx into focused modules (551→265 lines)
Extract session lifecycle (URL routing, popstate) into useChatSession hook
and SSE event handling (message assembly, tool progress) into useChatEvents
hook. ChatBar now only handles UI rendering and input.
